Update README and restructure demo app
- Reorganize demo files and update documentation
- Add service architecture details and usage instructions
- Update docker configuration with health checks
