+++ docker-compose.yml
@@ -5,16 +5,22 @@ services:
       - ${OLLAMA_PORT:-7869}:${OLLAMA_INTERNAL_PORT:-11434}
     volumes:
       - .:/code
-      - ./ollama/ollama:/root/.ollama
-    container_name: ollama
+      - ./.ollama:/root/.ollama
     pull_policy: always
     tty: true
     restart: always
     environment:
       - OLLAMA_KEEP_ALIVE=${OLLAMA_KEEP_ALIVE:-24h}
-      - OLLAMA_HOST=${OLLAMA_HOST:-0.0.0.0}
-    networks:
-      - ollama-docker
+      - OLLAMA_HOST=0.0.0.0
+    entrypoint: ["sh", "-c"]
+    command: ["ollama serve & sleep 10 && ollama pull ${INFERENCE_MODEL:-llama3.2:1b} && wait"]
+    healthcheck:
+      test: ["CMD", "sh", "-c", "ollama list | grep -q '${INFERENCE_MODEL:-llama3.2:1b}'"]
+      interval: 30s
+      timeout: 15s
+      retries: 40
+      start_period: 300s
+
 
   llama-stack:
     image: llamastack/distribution-ollama:latest
@@ -22,7 +28,6 @@ services:
       - ${LLAMA_STACK_PORT:-5001}:5000
     volumes:
       - ./.llama:/root/.llama
-    container_name: llama-stack
     pull_policy: always
     restart: always
     environment:
@@ -31,9 +36,8 @@ services:
       - LLAMA_STACK_PORT=5000
     command: --port 5000 --template ollama
     depends_on:
-      - ollama
-    networks:
-      - ollama-docker
+      ollama:
+        condition: service_healthy
     healthcheck:
       test: ["CMD", "curl", "-f", "http://localhost:5000/v1/models"]
       interval: 30s
@@ -49,22 +53,19 @@ services:
       - ${PLAYGROUND_PORT:-8501}:${PLAYGROUND_PORT:-8501}
     volumes:
       - ./.llama:/root/.llama
-    container_name: llama-stack-playground
     restart: always
     environment:
       - LLAMA_STACK_ENDPOINT=http://llama-stack:5000
     depends_on:
-      - llama-stack
-    networks:
-      - ollama-docker
+      llama-stack:
+        condition: service_healthy
 
-  chainlit:
+  chainlit-app:
     build:
       context: .
-      dockerfile: Dockerfile.chainlit
+      dockerfile: Dockerfile.demo_01
     ports:
-      - ${CHAINLIT_PORT:-8000}:8000
-    container_name: chainlit-app
+      - ${CHAINLIT_PORT:-9001}:8000
     restart: always
     environment:
       - LLAMA_STACK_API_URL=http://llama-stack:5000
@@ -72,9 +73,9 @@ services:
     depends_on:
       llama-stack:
         condition: service_healthy
-    networks:
-      - ollama-docker
-
-networks:
-  ollama-docker:
-    driver: bridge
+    healthcheck:
+      test: ["CMD", "curl", "-f", "http://localhost:8000"]
+      interval: 30s
+      timeout: 10s
+      retries: 5
+      start_period: 120s
